+++ Assets/StargateNet/StargateNet/Base/Config/StargateConfigData.cs
@@ -5,7 +5,7 @@ namespace StargateNet
 {
     public struct StargateConfigData
     {
-        public float tickRate;
+        public int tickRate;
         public ushort maxClientCount;
         public bool runAsHeadless;
         public List<GameObject> networkPrefabs;
